refector: 컴포넌트/css 분리 및 폴더 이동
