+++ app/(home)/page.module.scss
@@ -7,48 +7,6 @@ $border-color: #e8e8e8;
   display: flex; // Flexbox 레이아웃 사용
   flex-direction: column; // Flexbox 방향을 세로로 설정
 
-  .calendarHeader {
-    width: 100%;
-    font-size: 24px;
-    font-weight: 700;
-    height: 32px;
-    margin-bottom: 12px;
-    display: flex;
-
-    // Dropdown 컨테이너
-    & > :first-child {
-      flex: 1;
-      display: flex;
-      justify-content: left; // 드롭다운 중앙 정렬
-    }
-
-    .pagination {
-      flex: 1;
-      display: flex;
-      gap: 20px;
-      align-items: center;
-      justify-content: center; // 페이지네이션 중앙 정렬
-
-      .paginationIcon {
-        color: var(--gray);
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        cursor: pointer;
-
-        &:hover {
-          color: var(--foreground);
-        }
-      }
-    }
-    // 수업 추가 버튼 컨테이너
-    .classAddButton {
-      flex: 1;
-      display: flex;
-      justify-content: right; // 버튼 중앙 정렬
-    }
-  }
-
   .CalendarSection {
     background-color: var(--content); // 배경색 설정
     width: 100%; // 부모 요소의 너비를 100%로 설정
@@ -76,173 +34,5 @@ $border-color: #e8e8e8;
         font-size: 1.7rem;
       }
     }
-
-    .calendarDate {
-      flex-grow: 1; // 남은 공간을 모두 차지하도록 설정
-      display: flex; // Flexbox 레이아웃 사용
-      flex-direction: column; // Flexbox 방향을 세로로 설정
-    }
-
-    .calendarRow {
-      background-color: var(--background); // 배경색 설정
-      display: flex; // Flexbox 레이아웃 사용
-      height: calc(100% / 6); // 높이를 1/6로 설정
-
-      .dayComponent {
-        border: 1px solid $border-color;
-        flex: 1;
-        display: flex;
-        flex-direction: column;
-        // padding: 12px 12px 0 12px;
-
-        &.expanded {
-          z-index: 1;
-
-          .classList {
-            height: auto; // expanded 상태에서는 자동 높이
-            background-color: var(--background); // 배경색 설정
-            border-bottom: 4px solid $border-color;
-          }
-        }
-
-        .dayIndex {
-          margin: 12px 12px 0 12px;
-          font-size: 19px;
-          font-weight: 700;
-          line-height: 1; // 텍스트의 기본 라인 높이를 1로 설정하여 여백 없애기
-        }
-
-        .classList {
-          height: 75px; // 기본적으로 3개 수업까지만 보이도록 설정
-          margin-top: 3px;
-          overflow-y: inherit;
-
-          .classItem {
-            font-size: 1.3rem;
-            border: 2px solid black;
-            border-radius: 10px;
-            box-sizing: border-box;
-            padding: 0 6px;
-            height: 24px;
-            display: flex;
-            justify-content: space-between;
-            margin: 0 12px;
-
-            .userIconContainer {
-              display: none;
-            }
-            .adminIconContainer {
-              display: flex;
-              gap: 4px;
-
-              .icon {
-                display: none;
-                cursor: pointer;
-                color: #8b8b8b;
-
-                &:hover {
-                  color: #000000; /* 아이콘 hover 시 색상 */
-                }
-              }
-            }
-            &:hover {
-              .icon {
-                display: block;
-              }
-            }
-          }
-        }
-        .toggleButton {
-          all: unset;
-          cursor: pointer;
-          width: 100%;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          border-radius: 5px;
-          margin-top: 3px;
-
-          &:hover {
-            background-color: #bebcbc;
-          }
-        }
-      }
-
-      .otherMonth {
-        background-color: var(--content);
-        color: var(--gray);
-        &.expanded {
-          .classList {
-            background-color: var(--content); // 배경색 설정
-          }
-        }
-      }
-    }
-  }
-
-  .modal {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100vw;
-    height: 100vh;
-    background-color: rgba(0, 0, 0, 0.15);
-    display: none;
-    justify-content: center;
-    align-items: center;
-
-    &.modalOpen {
-      display: flex;
-    }
-    .modalContent {
-      width: 384px;
-      padding: 40px 70px;
-      background: var(--background);
-      color: black;
-      border-radius: 0.5rem;
-      display: flex;
-      flex-direction: column;
-
-      h1 {
-        font-size: 2rem;
-        font-weight: 700;
-        text-align: center;
-        margin-bottom: 20px;
-      }
-
-      .formGroup {
-        font-size: 1.5rem;
-        display: flex;
-        flex-direction: column;
-        margin-bottom: 10px;
-
-        label {
-          margin-bottom: 3px;
-          font-weight: 400;
-        }
-
-        .isWrong {
-          > * {
-            // isWrong의 직계 자식 요소 모두 선택
-            outline: 1px solid var(--red);
-            border-radius: 5px;
-            box-sizing: border-box;
-          }
-        }
-      }
-
-      .modalButton {
-        display: flex;
-        justify-content: center;
-        gap: 20px;
-        margin-top: 20px;
-      }
-    }
-  }
-
-  // admin이 아닐 때 기본적으로 숨김
-  .classAddButton,
-  .adminIconContainer {
-    visibility: visible;
   }
 }
