+++ components/classCalendar/calendarDate/calendarDate.module.scss
@@ -1,0 +1,13 @@
+$border-color: #e8e8e8;
+
+.calendarDate {
+  flex-grow: 1; // 남은 공간을 모두 차지하도록 설정
+  display: flex; // Flexbox 레이아웃 사용
+  flex-direction: column; // Flexbox 방향을 세로로 설정
+}
+
+.calendarRow {
+  background-color: var(--background); // 배경색 설정
+  display: flex; // Flexbox 레이아웃 사용
+  height: calc(100% / 6); // 높이를 1/6로 설정
+}
