+++ components/classCalendar/classItem/classItem.module.scss
@@ -1,0 +1,34 @@
+.classItem {
+  font-size: 1.3rem;
+  border: 2px solid black;
+  border-radius: 10px;
+  box-sizing: border-box;
+  padding: 0 6px;
+  height: 24px;
+  display: flex;
+  justify-content: space-between;
+  margin: 0 12px;
+
+  .userIconContainer {
+    display: none;
+  }
+  .adminIconContainer {
+    display: flex;
+    gap: 4px;
+
+    .icon {
+      display: none;
+      cursor: pointer;
+      color: #8b8b8b;
+
+      &:hover {
+        color: #000000; /* 아이콘 hover 시 색상 */
+      }
+    }
+  }
+  &:hover {
+    .icon {
+      display: block;
+    }
+  }
+}
